# api/api-docs/build/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Xpense web api
  description: Xpense web api.
servers:
  - url: 'http://localhost:7000/api'
  - url: 'https://dev.api.xpense.bulldozerinc.com/api'
paths:
...
  '/org/{orgId}/sheets/{sheetId}':
    get:
      summary: get sheet by id
      tags:
        - Organization
        - Sheet
      operationId: getSheetById
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Get sheet by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sheet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: delete sheet
      tags:
        - Organization
        - Sheet
      operationId: deleteSheet
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Sheet successfully delete.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
...
  '/org/{orgId}/exp-types':
    get:
      summary: Get list of expense types
      tags:
        - Expense types
        - Organization
      operationId: getExpenseTypes
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: List of expense types
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseTypes'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create new expense type
      tags:
        - Organization
        - Expense type
      operationId: createExpenseType
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      requestBody:
        description: Expense type DTO object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseType'
      responses:
        '201':
          description: Expense type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/exp-types/{expenseTypeId}':
    get:
      summary: Get expense type by id
      tags:
        - Expense types
        - Organization
      operationId: getExpenseType
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: expenseTypeId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Expense type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseType'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      summary: Update expense type
      tags:
        - Organization
        - Expense type
      operationId: updateExpenseType
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: expenseTypeId
          schema:
            type: string
          required: true
      requestBody:
        description: Expense type DTO object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseType'
      responses:
        '204':
          description: Expense type successfully updated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete expense type
      tags:
        - Organization
        - Expense type
      operationId: deleteExpenseType
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: expenseTypeId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Expense type successfully delete.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
...
  '/sheets/{sheetId}/attachment':
    post:
      summary: upload attachment to sheet
      tags:
        - Sheets
      parameters:
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      operationId: insertAttachment
      requestBody:
        description: Upload attatchment file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Upload'
      responses:
        '200':
          description: uploaded
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  '/sheets/{sheetId}/expenses':
    post:
      summary: create expense
      tags:
        - Expenses
      parameters:
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      operationId: createExpense
      requestBody:
        description: create expense body request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Expense'
      responses:
        '204':
          description: Expense created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    get:
      summary: Get expenses by sheet
      tags:
        - Expenses
      operationId: getExpenses
      parameters:
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Expenses list by sheet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expenses'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  '/sheets/{sheetId}/expenses/{expenseId}':
    get:
      summary: Get expense by id
      tags:
        - Expenses
      operationId: getExpense
      parameters:
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
        - in: path
          name: expenseId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update expense
      tags:
        - Expenses
      parameters:
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
        - in: path
          name: expenseId
          schema:
            type: string
          required: true
      operationId: updateExpense
      requestBody:
        description: Update expense body request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExpense'
      responses:
        '204':
          description: Expense updated
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: delete expense
      tags:
        - Expenses
      parameters:
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
        - in: path
          name: expenseId
          schema:
            type: string
          required: true
      operationId: deleteExpense
      responses:
        '200':
          description: Expense deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/sheets/{sheetId}/expenses/{expenseId}/attachment':
    post:
      summary: upload attachment to expense
      tags:
        - Expenses
      parameters:
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
        - in: path
          name: expenseId
          schema:
            type: string
          required: true
      operationId: insertAttachment
      requestBody:
        description: Upload attatchment file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Upload'
      responses:
        '200':
          description: uploaded
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Sheet:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        status:
          type: string
        creationDate:
          type: string
          format: date-time
        submittedOn:
          type: string
          format: date-time
        exported:
          type: boolean
        total:
          type: integer
...
    Expense:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        sheetId:
          type: string
        type:
          type: string
        source:
          type: string
        label:
          type: string
        payMethod:
          type: string
        category:
          type: string
        date:
          type: integer
          format: date-time
          example: '2020-12-15'
        amount:
          type: object
          properties:
            curr:
              type: string
            value:
              type: integer
              example: 10.5
            tax:
              type: object
              properties:
                id:
                  type: string
                amount:
                  type: integer
                  example: 10.5
        attachments:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              originalName:
                type: string
    Expenses:
      type: array
      items:
        $ref: '#/components/schemas/Expense'
    UpdateExpense:
      type: object
      properties:
        date:
          type: string
          format: date-time
        type:
          type: string
        source:
          type: string
        amount:
          type: integer
          example: 52.5
        currency:
          type: string
        taxId:
          type: string
        taxAmount:
          type: integer
          example: 7.2
    ExpenseTypes:
      type: array
      items:
        $ref: '#/components/schemas/ExpenseType'
    ExpenseType:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        type:
          type: string
        multiplicator:
          type: integer
          example: 0.25
        status:
          type: string
        i18n:
          type: string
        attachmentRequired:
          type: boolean
    CreateExpenseType:
      type: object
      properties:
        code:
          type: string
        type:
          type: string
        multiplicator:
          type: integer
          example: 0.25
        status:
          type: string
        i18n:
          type: string
        attachmentRequired:
          type: boolean
...
    Upload:
      type: object
      properties:
        file:
          type: string
          format: binary
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    NotFound:
      description: The specified resource was not found
    BadRequest:
      description: Bad request
security:
  - bearerAuth: []