# app/src/models/expenseType.ts
export default class ExpenseType {
    id?: string;
    code?: string;
    type?: string;
    multiplicator?: number;
    status?:string;
    attachmentRequired?:string;
    i18n?: string;
    orgId?: string;
}